Fix tests for Windows

+++ src/main/java/net/openhft/chronicle/queue/impl/single/SingleChronicleQueue.java
@@ -777,14 +777,20 @@ public class SingleChronicleQueue extends AbstractCloseable implements RollingCh
         writeLock.lock();
         try {
             int cycle = cycle();
-            int lastCycle = lastCycle();
-            while (lastCycle < cycle && lastCycle >= 0) {
+            for (int lastCycle = lastCycle(); lastCycle < cycle && lastCycle >= 0; lastCycle--) {
                 try (final SingleChronicleQueueStore store = this.pool.acquire(lastCycle, epoch(), false, null)) {
+                    // file not found.
+                    if (store == null)
+                        break;
                     if (store.writePosition() == 0 && !store.file().delete() && store.file().exists()) {
                         // couldn't delete? Let's try writing EOF
                         // if this blows up we should blow up too so don't catch anything
-                        store.writeEOFAndShrink(wireType.apply(store.bytes()), timeoutMS);
-                        lastCycle--;
+                        MappedBytes bytes = store.bytes();
+                        try {
+                            store.writeEOFAndShrink(wireType.apply(bytes), timeoutMS);
+                        } finally {
+                            bytes.releaseLast();
+                        }
                         continue;
                     }
                     break;
